+++ llvm/lib/CMakeLists.txt
@@ -3,6 +3,7 @@
 
 add_subdirectory(IR)
 add_subdirectory(FuzzMutate)
+add_subdirectory(InterfaceStub)
 add_subdirectory(IRReader)
 add_subdirectory(CodeGen)
 add_subdirectory(BinaryFormat)
