+++ llvm/lib/LLVMBuild.txt
@@ -30,6 +30,7 @@ subdirectories =
  FuzzMutate
  LineEditor
  Linker
+ InterfaceStub
  IR
  IRReader
  LTO
