+++ llvm/lib/TextAPI/CMakeLists.txt
@@ -1,6 +1,4 @@
 add_llvm_component_library(LLVMTextAPI
-  ELF/ELFStub.cpp
-  ELF/TBEHandler.cpp
   MachO/Architecture.cpp
   MachO/ArchitectureSet.cpp
   MachO/InterfaceFile.cpp
