+++ llvm/unittests/CMakeLists.txt
@@ -26,6 +26,7 @@ add_subdirectory(Demangle)
 add_subdirectory(ExecutionEngine)
 add_subdirectory(Frontend)
 add_subdirectory(FuzzMutate)
+add_subdirectory(InterfaceStub)
 add_subdirectory(IR)
 add_subdirectory(LineEditor)
 add_subdirectory(Linker)
