+++ llvm/unittests/InterfaceStub/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(LLVM_LINK_COMPONENTS
+  InterfaceStub
+)
+
+add_llvm_unittest(InterfaceStubTests
+  ELFYAMLTest.cpp
+)
+
+target_link_libraries(InterfaceStubTests PRIVATE LLVMTestingSupport)
